+++ ts/src/components/grid/grid-row.component.tsx
@@ -1,23 +1,31 @@
 import { useState, useEffect } from 'react';
-import { DataModel } from '../../data';
-import { GridColumnProps } from './types';
+import { Rosie } from '../../core';
+import { GridColumnProps, GridRowProps } from './types';
 import { GridCell } from './grid-cell.component';
 
-export function GridRow(props: { record: DataModel<any>, rowIndex: number, columns: GridColumnProps[], checkboxSelection?: boolean, onSelectionChange?: (record: DataModel<any>) => void }) {
-  const { record, rowIndex, columns, checkboxSelection, onSelectionChange } = props,
+export function GridRow(props: GridRowProps) {
+  const { record, rowIndex, columns, checkboxSelection } = props,
         [selected, setSelected] = useState(false);
 
   useEffect(() => {
-    record.selected = selected;
-    onSelectionChange && onSelectionChange(record);
-  }, [selected])
+    if (record) {
+      record.on('select', () => {
+        setSelected(record?.selected ?? false);
+      });
+      record.trigger('select');
+    }
+  }, [record])
 
-  return <div key={rowIndex} className="rosie-grid-row d-flex flex-row">
-    {checkboxSelection && <div className="rosie-grid-cell p-2">
-      <div className="form-check mb-0">
-        <input className="form-check-input" type="checkbox" checked={selected} onChange={() => { setSelected(!selected); }} />
-      </div>
-    </div>}
-    {columns.map((col: GridColumnProps, colIndex) => <GridCell key={colIndex} {...col} record={record} rowIndex={rowIndex} colIndex={colIndex} />)}
-  </div>
+  return <>
+    <div className={Rosie.classNames('rosie-grid-row d-flex flex-row', { selected })}>
+      {checkboxSelection && <>
+        <div className="rosie-grid-cell p-1">
+          <div className="form-check mb-0">
+            <input className="form-check-input" type="checkbox" checked={selected} onChange={() => { record.selected = !selected; }} />
+          </div>
+        </div>
+      </>}
+      {columns.map((col: GridColumnProps, colIndex) => <GridCell key={colIndex} record={record} rowIndex={rowIndex} colIndex={colIndex} {...col} />)}
+    </div>
+  </>
 }