+++ ts/src/components/grid/grid.component.tsx
@@ -13,59 +13,78 @@ export function Grid(props: GridProps) {
         [columns, setColumns] = useState([] as GridColumnProps[]),
         [currentPage, setCurrentPage] = useState(1),
         [allSelected, setAllSelected] = useState(false),
-        [data, setData] = useState<DataModel<any>[]>([]);
+        [records, setRecords] = useState<DataModel<any>[]>([]),
+        [rawData, setRawData] = useState<DataModel<any>[]>([]);
 
   useEffect(() => {
-    const store$ = props.store?.subscribe(value => setData(value || []));
+    const store$ = props.store?.subscribe(value => setRawData(value || []));
 
     const body = document.querySelector(`#${gridId} .rosie-grid-body`);
-    body.addEventListener('scroll', (event: Event) => {
+    body.addEventListener('scroll', () => {
       document.querySelector(`#${gridId} .rosie-grid-header`).scrollLeft = body.scrollLeft;
     });
 
     return () => { store$?.unsubscribe(); }
   }, [])
 
-  useEffect(() => { !props.store && setData((props.data ?? []).map(DataModel.create)) }, [props.data])
+  useEffect(() => { !props.store && setRawData((props.data ?? []).map(DataModel.create)) }, [props.data])
 
   useEffect(() => {
     const columns = Children.toArray(props.children).map((child: any) => child.props as GridColumnProps);
     setColumns(columns);
   }, [props.children])
 
-  useEffect(() => { data.forEach(record => { record.selected = allSelected; }) }, [allSelected])
+  useEffect(() => { records.forEach(record => { record.selected = allSelected; }) }, [allSelected])
 
-  function getDisplayRows() {
-    if (!props.pagingToolbar) return data;
+  useEffect(() => {
+    if (!rawData?.length || !currentPage) return;
 
-    const { pageSize = 25 } = props.pagingToolbar;
-    return data.take(pageSize, pageSize * (currentPage - 1));
-  }
+    if (!props.pagingToolbar) {
+      setRecords([...rawData]);
+    } else {
+      const { pageSize = 25 } = props.pagingToolbar,
+            data = rawData.take(pageSize, pageSize * (currentPage - 1));
+      setRecords([...data]);
+    }
+  }, [rawData, currentPage])
 
-  return <div id={gridId} className={Rosie.classNames('rosie-grid rosie-grid-bordered rosie-grid-striped rosie-grid-hover d-flex flex-row', { fullscreen: props.fitScreen || props.fitHeight }, props.className)}>
-    <div className="rosie-grid-viewport d-flex flex-column fullscreen">
-      <div className={Rosie.classNames('rosie-grid-header overflow-hidden fw-bold bg-light d-flex', { 'flex-column': props.fitScreen || props.fitWidth })}>
-        <div className="rosie-grid-row d-flex flex-row">
-          {props.checkboxSelection && <div className="rosie-grid-cell p-2">
-          <div className="form-check mb-0"><input className="form-check-input" type="checkbox" checked={allSelected} onChange={() => setAllSelected(!allSelected)} /></div>
-          </div>}
-          {columns.map((col: GridColumnProps, index) => <GridCell header key={index} {...col} />)}
-          <div style={{width:Rosie.SCROLLBAR_WIDTH}} />
+  return <>
+    <div id={gridId} className={Rosie.classNames('rosie-grid d-flex flex-row', { fullscreen: props.fitScreen || props.fitHeight }, props.className)}>
+      <div className="rosie-grid-viewport d-flex flex-column fullscreen">
+        <div className={Rosie.classNames('rosie-grid-header overflow-hidden fw-bold bg-light d-flex', { 'flex-column': props.fitScreen || props.fitWidth })}>
+          <div className="rosie-grid-row d-flex flex-row">
+            {props.checkboxSelection && <>
+              <div className="rosie-grid-cell p-1">
+                <div className="form-check mb-0">
+                  <input className="form-check-input" type="checkbox" checked={allSelected} onChange={() => setAllSelected(!allSelected)} />
+                </div>
+              </div>
+            </>}
+            {/* {props.checkboxSelection && <div className="rosie-grid-cell p-2">
+              <div className="form-check mb-0"><input className="form-check-input" type="checkbox" checked={allSelected} onChange={() => setAllSelected(!allSelected)} /></div>
+            </div>} */}
+            {columns.map((col: GridColumnProps, index) => <GridCell header key={index} {...col} />)}
+            <div style={{width:Rosie.SCROLLBAR_WIDTH}} />
+          </div>
         </div>
+        <div className={Rosie.classNames('rosie-grid-body fullscreen overflow-x-auto d-flex', { 'flex-column': props.fitScreen || props.fitWidth, 'overflow-y-scroll': !props.fitWidth })}>
+          <div>
+            {(!records?.length) && <div className="p-2">No record found.</div>}
+            {(records?.length > 0) && records.map((record, rowIndex) => {
+              return <GridRow key={rowIndex} record={record} rowIndex={rowIndex} columns={columns} checkboxSelection={props.checkboxSelection} />
+            })}
+          </div>
+        </div>
+        {(props.bbar || props.pagingToolbar) && <>
+          <div className="rosie-grid-footer bg-light border-top d-flex flex-row p-2">
+            {props.pagingToolbar && <>
+              <div className="ms-auto">
+                <PagingToolbar data={rawData} onChange={setCurrentPage} />
+              </div>
+            </>}
+          </div>
+        </>}
       </div>
-      <div className={Rosie.classNames('rosie-grid-body fullscreen overflow-x-auto d-flex', { 'flex-column': props.fitScreen || props.fitWidth, 'overflow-y-scroll': !props.fitWidth })}>
-        <div>
-          {(!data || !data.length) && <div className="p-2">No record found.</div>}
-          {data?.length > 0 && getDisplayRows().map((record, rowIndex) => {
-            return <GridRow key={rowIndex} record={record} rowIndex={rowIndex} columns={columns} checkboxSelection={props.checkboxSelection} onSelectionChange={props.onSelectionChange} />
-          })}
-        </div>
-      </div>
-      {props.pagingToolbar && <div className="rosie-grid-footer bg-light border-top d-flex flex-row p-2">
-        <div className="ms-auto">
-          <PagingToolbar data={data} onChange={setCurrentPage} />
-        </div>
-      </div>}
     </div>
-  </div>
+  </>
 }